refactor(D2): extract FileMover into engine/file_mover.py
FileMoveTask and FileMover (~430 lines) are removed from data_logger.py
and live in their own module.  ScanManager creates the FileMover, sets
scan_number/save_local, and injects it via start_logging(file_mover=...).
ScanManager now accesses self.file_mover directly instead of chaining
through self.data_logger.file_mover.*

D1 safety-net tests (18) and all 61 engine tests pass.
